meta-nxp3220: check u-boot workdir status to clean build
This patch check u-boot workdir status to clean build

Reviewed-on: https://review.gerrithub.io/c/NexellCorp/yocto_meta-nexell/+/450410

+++ meta-nxp3220/recipes-bsp/u-boot/u-boot-2018.5-nexell.bb
@@ -28,7 +28,7 @@ do_configure_prepend () {
 	oe_runmake -C ${S} distclean
     else
         file=${B}/.uboot_defconfig
-	if [ -e ${file} ]; then
+	if [ -e ${file} ] && [ -d ${WORKDIR} ]; then
 	    conf=$(cat ${file})
 	    if [ "${conf}" != "${UBOOT_MACHINE}" ]; then
 		rm ${file}; echo ${UBOOT_MACHINE} >> ${file};
